requirements.txt
scrapy
pymongo
requests
python-dotenv